backend/requirements.txt
ffmpeg-python==0.2.0
Pillow>=9.1.0
pydantic>=2.9.2
fastapi>=0.115.0
uvicorn>=0.31.0
psutil>=6.0.0